backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pandas==2.1.3
openai==1.3.7
groq==0.11.0
google-generativeai==0.8.3
plotly==5.24.1
numpy==1.26.4
scipy==1.11.4
python-dotenv==1.0.0
httpx==0.27.2
openpyxl==3.1.2
